Bail immediately during rb-site if the Django environment can't load.
When installing Review Board, a bad dependency or other issue
initializing Django would result in a breakage that won't be noticed
until later in the process. This would result in a series of
installation steps that say ERROR, but continue on until an exception
was raised.

We now check for errors here, and immediately quit with a helpful
message so it doesn't appear that the database was created or other
steps were taken.

Testing Done:
Hit this while doing some testing and got very confused trying to
install a site.

Fixed it with this patch and repeated the failed install. Saw the
error message instead of a supposed "successful installation"
followed by traceback.

Reviewed at https://reviews.reviewboard.org/r/14538/