Fix unit test for the new css_media and js_media context data.
During the development of the change to include css_media and js_media
in the data for the files in the diff viewer for binary files which use
review UIs, these keys got moved from the diffviewer code into reviews,
which creates its own list of files. The file list created by the
reviews code goes into a separate key in the context, and the unit test
for these wasn't updated, causing the test to fail. This change fixes
that.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14594/