Add typing and fixes for reviewboard.diffviewer.commit_utils.
This fleshes out the remainder of reviewboard.diffviewer.commit_utils
with typing and documentation. The validation payloads are now fully
typed, which is only useful internally as this is considered opaque
outside this module.

There is one fix with equality checking for CommitHistoryDiffEntry. It
assumed the other type, and would crash if given a different type. We
now check for this.

This is mostly the same as Christian's change at /r/14416. The changes
I've made:

- Moved `SerializedCommitHistoryDiffEntryType` so it's defined before
  it's used.
- Fixed a typo in the `DiffCommitsValidationInfo` docs.
- Fixed up some Args/Returns types to list the specific
  TypedDict/TypeAlias names instead of just `str` or `dict`.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14596/