Update unit tests for diff line, opcode, and chunk type changes.
Diff lines, opcodes, and chunks are now all explicitly typed, and we
made a few tweaks for correctness (such as using `None` for the line
numbers for the "other" side of an inserted/deleted chunk instead of an
empty string).

This change updates the unit tests that touch these types to make sure
we're always fully defining the structures. Many of these tests used
partial dictionaries for chunks, or even used dictionaries with integer
keys instead of tuples for lines. This of course was flagged by the type
checker.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14595/