Deprecate Review Board's assertQueries() method.
The underlying Djblets TestCase.assertQueries() has had a deprecation
warning, but Review Board didn't. This still triggered warnings, but the
line number associated with those made it harder to find the actual
instances, since it'd just show the call to `super().assertQueries()`.
This change marks the subclass version for deprecation at the matching
version to when we'll be removing the Djblets one.

This also cleans up a little bit of adjacent code that was very out of
date.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14603/