Fix ImageMimetypeTests to call super().setUp()
The `ImageMimetypeTests` class has a `setUp()` method, but it didn't
call the parent class one. I discovered this while running tests via
pytest-xdist, and hit failures in this class because it was the first
set of tests running on that thread, and `initialize()` had not been
called before we were trying to access the review UI registry.

Testing Done:
Ran unit tests with xdist and a number of threads that exposed this bug,
saw everything succeed.

Reviewed at https://reviews.reviewboard.org/r/14616/