Tree Sitter: Add new dependencies.
This change adds the new dependencies for `tree-sitter` and
`tree-sitter-language-pack`.

These are pinned to exact versions. `tree-sitter`'s API is still in a bit
of flux, and they shipped some breaking changes between 0.24 and 0.25,
so I'm not confident in specifying any kind of range. For
`tree-sitter-language-pack`, we'll probably always need to pin an exact
version, because the queries and the grammars need to be compatible or
else the queries won't load at all.

Note that there is one bug fix that I submitted for the tree-sitter
package which has not yet been released. I'll bump the dependency once
that happens.

Testing Done:
Ran setup.py develop.

Reviewed at https://reviews.reviewboard.org/r/14512/