Tree Sitter: Add update-language-info script and resulting data.
This change adds a script that loads information about supported
languages from a tree-sitter-language-pack checkout. This creates a
file, `reviewboard/treesitter/_languages.py`, which has two mappings,
`MIME_TYPE_TO_LANGUAGE` and `FILE_SUFFIX_TO_LANGUAGES`, which are used
for determining the tree sitter language name to use for a given file.

The compiled grammars that are shipped in tree-sitter-language-pack
don't include any of the grammar metadata, so this script will go into a
a checkout and grab the filename suffixes. This also fetches the
"first_line_regex" key, although that is not currently used for
detection.

Testing Done:
Ran script to create the _languages.py file.

Reviewed at https://reviews.reviewboard.org/r/14514/