Tree Sitter: Use Tree Sitter for highlighting in the chunk generator.
This change makes it so we attempt to use tree sitter (if available) for
diffs. If the tree sitter highlighting is unavailable (for example, we
don't have a matching grammar), or fails for some reason, we'll then
fall back on Pygments.

Testing Done:
- Ran unit tests.
- Viewed a variety of diffs of different file types and verified that
  the new highlighting worked correctly.

Reviewed at https://reviews.reviewboard.org/r/14524/