Port the Read Only Banner to the new page banners.
The built-in Read Only Banner is now built on top of the new page banner
component, helping it fit in with any other banners on display. This
greatly simplifies the code that needs to be maintained as well.

The banner is always shown above all other banners.

Testing Done:
Tested the site in read-only mode on desktop and mobile, with and
without other banners being shown. Verified that it appeared correctly.

Reviewed at https://reviews.reviewboard.org/r/14559/