Support session tokens and non-JSON responses for license checks.
When preparing for a license check on the Licenses page, a License
Provider can now provide a Session Token, which is an opaque value that
will be passed back during processing of a response. That allows
information to be temporarily stored and then retrieved for response
processing.

Responses are also no longer required to be JSON. A license server may
now provide any kind of text-based response, which may be processed by
the License Provider however needed.

Testing Done:
Unit tests pass.

Tested this with the in-progress work in Power Pack.

Reviewed at https://reviews.reviewboard.org/r/14613/