Fix a missing import in the licensing code due to a merge.
While working on the licensing, a big change was landed that moved
various types over to `typelets`. When updating the licensing code, a
bad merge resulted in an import getting left in a `TYPE_CHECKING` block
instead of being promoted to a top-level import as intended.

This change moves the import.

Testing Done:
Unit tests pass.

Verified this fixed a crash when using license providers.

Reviewed at https://reviews.reviewboard.org/r/14625/