Fix a bad style in CodeMirror.
We had a weird, single style rule for `.cm-header` that just specified
the color as "black". This made markdown header lines (such as those
starting with `#`) very hard to read when running in dark mode.

This change gets rid of that.

Testing Done:
Edited a change description that had header lines and was able to read
the text.

Reviewed at https://reviews.reviewboard.org/r/14669/