Fix processing of comments when there's no Review UI.
When there's no Review UI for a file attachment, the thumbnail view
itself allows adding a comment for a file. This was processing through
the comments data expecting those to be an array, but after the work to
add typing and make all the individual review UIs consistent, the data
here is an object with arrays as the values. This change fixes it to
properly iterate over that.

Testing Done:
Was able to add a new comment on a file attachment that did not have a
review UI.

Reviewed at https://reviews.reviewboard.org/r/14657/