Fix table display of selected users in related user selector.
The way that selected users were appearing in the related user selector
was kind of ugly. This change makes it so the third table cell expands,
pushing the user avatar and username to the left.

This also fixes the display of the placeholder text.

Testing Done:
Used the related user selector and saw that things looked nice now.

Reviewed at https://reviews.reviewboard.org/r/14670/