Use the correct method to get a license comm request's headers.
When checking on the status of a license, we attempt to get the
`Content-Type` header of a license request. This was done by accessing
the header as a dictionary lookup, but this isn't standard. Instead, we
should use the `.get()` method instead. This updates to use that.

Testing Done:
Unit tests pass.

Manually tested license updates and communication.

Reviewed at https://reviews.reviewboard.org/r/14655/