Speed up looking up actions by attachment point.
When populating the UI, we fetch actions based on various attachment
points. When we do this, the registry has to iterate through every
action every time, returning each one matching the attachment point.
This is algorithmically slow, and only gets worse as we expand usage of
actions.

Now, when registering an action, the action gets placed in a cache of
attachment points to actions. Unregistering will remove that action from
the cache. This makes lookup very fast.

Testing Done:
Unit tests pass.

Tested various parts of the UI, making sure the right actions are in
the right places.

Reviewed at https://reviews.reviewboard.org/r/14639/