Add a concept of verbose labels to actions.
A verbose label is a longer version of a label that can be used for
display purposes (such as in the Review menu) and for ARIA labels. By
separating these concepts, an action can have one representation in a
place like the Review menu and another in a place like the Quick Access
area without having to have two separate registered actions.

It's set up just like labels on the Python classes and JavaScript
model, and is opted into by action views.

The detailed menu template (used for the Review menu) prioritizes the
verbose label over the standard label. Standard menus and buttons use it
exclusively for ARIA labels.

Testing Done:
Unit tests pass.

Tested that the Review menu had the correct labels and that buttons had
the correct ARIA labels.

Reviewed at https://reviews.reviewboard.org/r/14642/