Provide the action's ID to the JavaScript model as "id".
When actions were first implemented, the action's ID was provided to
the model as an `actionId` attribute. This wasn't marked as the model's
attribute, so it didn't properly function as an ID for lookup purposes.

To keep things clean and standard, this has been renamed to `id`. This
may have some theoretical fallout if anyone has actions with custom
views that reference `this.get('actionId')`, but it's unlikely any
third-parties will truly be affected. If needed, we could alias the ID.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/14661/